+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Changelog
 
+##### not released
+
+    * Code clean up
+    * Unregister index before closing
+
+##### hugo-search 1.0.2 - (08 September 2016)
+
+	* Fix build due to API changes
+
 ##### hugo-search 1.0.1 - (09 August 2016)
 
     * Fix small bug in search.js